перенос функций SendRequest и request
