+++ request.js
@@ -1,0 +1,42 @@
+/* jshint esversion: 6 */
+
+const
+    { stringify } = require('querystring'),
+    { request } = require('https');
+
+module.exports = {
+    send: (params, postData) => {
+        return new Promise((resolve, reject) => {
+            let req = request(params, (res) => {
+                res.setEncoding('utf8');
+                res.on('data', (data) => resolve( JSON.parse(data) ));
+            });
+            
+            req.on('error', reject);
+            
+            if(postData) req.write(postData);
+            req.end();
+        });
+    },
+    request: (params) => {
+        if(params.body) {
+            let postData = stringify( String(params.body) );
+            params.headers['Content-Type'] = 'application/x-www-form-urlencoded';
+            params.headers['Content-Length'] = Buffer.byteLength(postData);
+    
+            return send(params, postData);
+        } else return send(params);
+    },
+    sendStat: (client, opt) => {
+        let shards = 0;
+        if (client.shard && client.shard.count !== 1) shards = client.shard.count;
+        else if (client.shards && client.shards.size !== 1) shards = client.shards.size;
+
+        opt.body = { servers: client.guilds.size, shards: shards };
+        return request(opt)
+            .then((r) => {
+                if(r.error) return console.error("[sdc-api | Авто-пост] Ошибка в работе\n" + r.error.message);
+                else return console.info("[sdc-api | Авто-пост] Статистика для " + client.user.tag + " опубликована на мониторинг.\n" + encodeURI(paths.botsPath + "/" + client.user.id));
+            }, (e) => console.error("[sdc-api | Авто-пост] Ошибка в работе | ", e));
+    }
+};
